flask-api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.1.0
pytesseract==0.3.10
gunicorn==21.2.0